Tabs: Move tabs into widgets directory
